backend/requirements.txt
# Core Django packages
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.1.0

# Database
psycopg2-binary==2.9.7

# API Documentation
drf-yasg==1.21.7

# Image processing
Pillow==9.5.0

# Testing and development
faker==18.13.0
pytest==7.3.1
pytest-django==4.5.2
coverage==7.2.7

# Utilities
python-dateutil==2.8.2
requests==2.30.0python-decouple==3.8
